TIP-1298 Moves the query to right adapter
