+++ tests/back/Platform/Acceptance/Analytics/Resources/config/queries.yml
@@ -1,3 +1,3 @@
 services:
     pim_analytics.query.email_domains:
-        class: AkeneoTest\Platform\Acceptance\Analytics\Query\EmailDomainsInMemory
+        class: AkeneoTest\Platform\Acceptance\Analytics\Query\Sql\EmailDomainsInMemory
