+++ tests/back/Platform/Integration/Analytics/Query/Sql/EmailDomainsIntegration.php
